requirements.txt
allure-pytest==2.8.22
attrs==20.3.0
cattrs==1.1.2
cx-Oracle==8.0.1
Faker==5.0.2
pytest==6.1.2
pytest-testrail==2.9.0
requests==2.25.0
sqlacodegen==2.3.0
sqlalchemy==1.4.20
xmltodict==0.12.0
httpx